+++ src/whispercpp/api_cpp2py_export.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #ifdef BAZEL_BUILD
 #include "context.h"
 #include "examples/common.h"
@@ -13,10 +15,6 @@
 #include "pybind11/pybind11.h"
 #include "pybind11/stl.h"
 #endif
-#include <algorithm>
-#include <functional>
-#include <numeric>
-#include <sstream>
 
 namespace py = pybind11;
